Refactor action activation tests to test db polling.
- Lowers max wait time for tests (shaves couple of minutes).
- Removes duplicate tests.
