+++ common/scala/src/main/resources/application.conf
@@ -438,7 +438,8 @@ whisk {
 
     controller {
         activation {
-            polling-from-db: true
+            polling-from-db = true
+            max-wait-for-blocking-activation = 60 seconds
         }
     }
