+++ tests/src/test/scala/org/apache/openwhisk/core/controller/test/ActionsApiTests.scala
@@ -1431,63 +1431,7 @@ class ActionsApiTests extends ControllerTestCommon with WhiskActionsApi {
     }
   }
 
-  it should "invoke an action, blocking with default timeout" in {
-    implicit val tid = transid()
-    val action = WhiskAction(
-      namespace,
-      aname(),
-      jsDefault("??"),
-      limits = ActionLimits(TimeLimit(1 second), MemoryLimit(), LogLimit()))
-    put(entityStore, action)
-    Post(s"$collectionPath/${action.name}?blocking=true") ~> Route.seal(routes(creds)) ~> check {
-      // status should be accepted because there is no active ack response and
-      // db polling will fail since there is no record of the activation
-      status should be(Accepted)
-      val response = responseAs[JsObject]
-      response.fields("activationId") should not be None
-      headers should contain(RawHeader(ActivationIdHeader, response.fields("activationId").convertTo[String]))
-    }
-  }
-
-  if (controllerActivationConfig.pollingFromDb) {
-    it should "invoke an action, blocking and retrieve result via db polling" in {
-      implicit val tid = transid()
-      val action = WhiskAction(namespace, aname(), jsDefault("??"))
-      val activation = WhiskActivation(
-        action.namespace,
-        action.name,
-        creds.subject,
-        activationIdFactory.make(),
-        start = Instant.now,
-        end = Instant.now,
-        response = ActivationResponse.success(Some(JsObject("test" -> "yes".toJson))),
-        logs = ActivationLogs(Vector("first line", "second line")))
-      put(entityStore, action)
-      // storing the activation in the db will allow the db polling to retrieve it
-      // the test harness makes sure the activation id observed by the test matches
-      // the one generated by the api handler
-      storeActivation(activation, context)
-      try {
-        Post(s"$collectionPath/${action.name}?blocking=true") ~> Route.seal(routes(creds)) ~> check {
-          status should be(OK)
-          val response = responseAs[JsObject]
-          response should be(activation.withoutLogs.toExtendedJson())
-        }
-
-        // repeat invoke, get only result back
-        Post(s"$collectionPath/${action.name}?blocking=true&result=true") ~> Route.seal(routes(creds)) ~> check {
-          status should be(OK)
-          val response = responseAs[JsObject]
-          response should be(activation.resultAsJson)
-          headers should contain(RawHeader(ActivationIdHeader, activation.activationId.asString))
-        }
-      } finally {
-        deleteActivation(ActivationId(activation.docid.asString), context)
-      }
-    }
-  }
-
-  it should "invoke an action, blocking and retrieve result via active ack" in {
+  it should "invoke a blocking action and retrieve result via active ack" in {
     implicit val tid = transid()
     val action = WhiskAction(namespace, aname(), jsDefault("??"))
     val activation = WhiskActivation(
@@ -1502,7 +1446,7 @@ class ActionsApiTests extends ControllerTestCommon with WhiskActionsApi {
 
     try {
       // do not store the activation in the db, instead register it as the response to generate on active ack
-      loadBalancer.whiskActivationStub = Some((1.milliseconds, activation))
+      loadBalancer.whiskActivationStub = Some((1.milliseconds, Right(activation)))
 
       Post(s"$collectionPath/${action.name}?blocking=true") ~> Route.seal(routes(creds)) ~> check {
         status should be(OK)
@@ -1521,7 +1465,7 @@ class ActionsApiTests extends ControllerTestCommon with WhiskActionsApi {
     }
   }
 
-  it should "invoke an action, blocking up to specified timeout and retrieve result via active ack" in {
+  it should "invoke a blocking action, waiting up to specified timeout and retrieve result via active ack" in {
     implicit val tid = transid()
     val action = WhiskAction(namespace, aname(), jsDefault("??"))
     val activation = WhiskActivation(
@@ -1536,16 +1480,18 @@ class ActionsApiTests extends ControllerTestCommon with WhiskActionsApi {
 
     try {
       // do not store the activation in the db, instead register it as the response to generate on active ack
-      loadBalancer.whiskActivationStub = Some((300.milliseconds, activation))
+      loadBalancer.whiskActivationStub = Some((300.milliseconds, Right(activation)))
 
       Post(s"$collectionPath/${action.name}?blocking=true&timeout=0") ~> Route.seal(routes(creds)) ~> check {
         status shouldBe BadRequest
-        responseAs[String] should include(Messages.invalidTimeout(WhiskActionsApi.maxWaitForBlockingActivation))
+        responseAs[String] should include(
+          Messages.invalidTimeout(controllerActivationConfig.maxWaitForBlockingActivation))
       }
 
       Post(s"$collectionPath/${action.name}?blocking=true&timeout=65000") ~> Route.seal(routes(creds)) ~> check {
         status shouldBe BadRequest
-        responseAs[String] should include(Messages.invalidTimeout(WhiskActionsApi.maxWaitForBlockingActivation))
+        responseAs[String] should include(
+          Messages.invalidTimeout(controllerActivationConfig.maxWaitForBlockingActivation))
       }
 
       // will not wait long enough should get accepted status
@@ -1553,7 +1499,6 @@ class ActionsApiTests extends ControllerTestCommon with WhiskActionsApi {
         val response = responseAs[JsObject]
         status shouldBe Accepted
         headers should contain(RawHeader(ActivationIdHeader, response.fields("activationId").convertTo[String]))
-
       }
 
       // repeat this time wait longer than active ack delay
@@ -1565,36 +1510,6 @@ class ActionsApiTests extends ControllerTestCommon with WhiskActionsApi {
       }
     } finally {
       loadBalancer.whiskActivationStub = None
-    }
-  }
-
-  if (controllerActivationConfig.pollingFromDb) {
-    it should "invoke a blocking action and return error response when activation fails" in {
-      implicit val tid = transid()
-      val action = WhiskAction(namespace, aname(), jsDefault("??"))
-      val activation = WhiskActivation(
-        action.namespace,
-        action.name,
-        creds.subject,
-        activationIdFactory.make(),
-        start = Instant.now,
-        end = Instant.now,
-        response = ActivationResponse.whiskError("test"))
-      put(entityStore, action)
-      // storing the activation in the db will allow the db polling to retrieve it
-      // the test harness makes sure the activation id observed by the test matches
-      // the one generated by the api handler
-      storeActivation(activation, context)
-      try {
-        Post(s"$collectionPath/${action.name}?blocking=true") ~> Route.seal(routes(creds)) ~> check {
-          status should be(InternalServerError)
-          val response = responseAs[JsObject]
-          response should be(activation.withoutLogs.toExtendedJson())
-          headers should contain(RawHeader(ActivationIdHeader, response.fields("activationId").convertTo[String]))
-        }
-      } finally {
-        deleteActivation(ActivationId(activation.docid.asString), context)
-      }
     }
   }
